reorganize module + some clippy fix
