+++ examples/pong/src/main.rs
@@ -1,28 +1,19 @@
 extern crate zengine;
 
-use zengine::basic::input::Bindings;
-use zengine::basic::input::InputType;
-use zengine::basic::input::{InputHandler, InputSystem};
-use zengine::basic::platform::PlatformSystem;
-use zengine::basic::render::Background;
-use zengine::basic::render::WindowSpecs;
-use zengine::basic::render::{RenderSystem, Sprite};
-use zengine::basic::timing::{FrameLimiter, TimingSystem};
 use zengine::core::system::Read;
-use zengine::core::system::{ReadSet, WriteSet};
-use zengine::core::Component;
-use zengine::core::Scene;
-use zengine::core::Store;
-use zengine::core::System;
-use zengine::core::Trans;
-use zengine::event::event_stream::EventStream;
-use zengine::event::event_stream::SubscriptionToken;
+use zengine::core::system::ReadSet;
+use zengine::core::system::WriteSet;
+use zengine::core::*;
+use zengine::event::*;
 use zengine::graphics::color::Color;
 use zengine::log::{info, trace, LevelFilter};
 use zengine::math::transform::Transform;
 use zengine::math::vector3::Vector3;
+use zengine::platform::*;
+use zengine::render::*;
 use zengine::serde::Deserialize;
 use zengine::serde_yaml;
+use zengine::timing::*;
 use zengine::Engine;
 
 fn main() {
@@ -112,11 +103,11 @@ impl Scene for Game {
             .build();
     }
 
-    fn on_stop(&mut self, store: &mut Store) {
+    fn on_stop(&mut self, _store: &mut Store) {
         trace!("Game scene on stop");
     }
 
-    fn update(&mut self, store: &Store) -> Trans {
+    fn update(&mut self, _store: &Store) -> Trans {
         match self.execution_numer {
             0 => Trans::None,
             _ => {
@@ -160,7 +151,7 @@ impl<'a> System<'a> for System1 {
         Read<'a, InputHandler<UserInput>>,
     );
 
-    fn init(&mut self, store: &mut Store) {
+    fn init(&mut self, _store: &mut Store) {
         trace!("setup system 1");
     }
 
@@ -182,7 +173,7 @@ impl<'a> System<'a> for System1 {
         self.run_count += 1;
     }
 
-    fn dispose(&mut self, store: &mut Store) {
+    fn dispose(&mut self, _store: &mut Store) {
         trace!("dispose system 1");
     }
 }
@@ -207,7 +198,7 @@ impl<'a> System<'a> for System2 {
         );
     }
 
-    fn run(&mut self, (set, stream): Self::Data) {
+    fn run(&mut self, (set, _stream): Self::Data) {
         trace!("run {} system 2", self.run_count);
 
         trace!("data 2: {:?}", set);
@@ -216,7 +207,7 @@ impl<'a> System<'a> for System2 {
         self.run_count += 1;
     }
 
-    fn dispose(&mut self, store: &mut Store) {
+    fn dispose(&mut self, _store: &mut Store) {
         trace!("dispose system 2");
     }
 }
