--- zengine/src/basic/platform/mod.rs
@@ -1,4 +1,0 @@
-mod platform;
-pub mod resources;
-
-pub use platform::PlatformSystem;
