+++ zengine/src/core/store.rs
@@ -2,8 +2,8 @@ use crate::core::component::Component;
 use crate::core::component::Components;
 use crate::core::component::Set;
 use crate::core::entity::{Entities, Entity, EntityBuilder};
-use crate::event::event_stream::EventStream;
-use crate::event::event_stream::SubscriptionToken;
+use crate::event::EventStream;
+use crate::event::SubscriptionToken;
 use downcast_rs::Downcast;
 use fnv::FnvHashMap;
 use std::any::Any;
