+++ zengine/src/platform/mod.rs
@@ -1,0 +1,9 @@
+extern crate sdl2;
+
+use crate::core::Resource;
+use sdl2::VideoSubsystem;
+
+mod system;
+pub use self::system::PlatformSystem;
+
+impl Resource for VideoSubsystem {}
